классы helpers для мапинга функций активации, for example

+++ ClassificationNumbers/MainClasses/Relation.cs
@@ -2,11 +2,11 @@
 {
     public class Relation
     {
-        private float _weight;
+        private double _weight;
         private Neuron _inputNeuron;
         private Neuron _outputNeuron;
 
-        public Relation(Neuron inputNeuron, Neuron outputNeuron, float weight)
+        public Relation(Neuron inputNeuron, Neuron outputNeuron, double weight)
         {
             _inputNeuron = inputNeuron;
             _outputNeuron = outputNeuron;
